+++ src/components/textbox/__internal__/__next__/text-input.component.tsx
@@ -24,6 +24,7 @@ import Label from "../../../../__internal__/label";
 import useRegisterValidationToTabs from "../../../../hooks/__internal__/useRegisterValidationToTabs/useRegisterValidationToTabs";
 import combineRefs from "../../../../__internal__/utils/helpers/combine-refs";
 import FieldsetValidationContext from "../../../../__internal__/fieldset-validation-context";
+import FieldsetContext from "../../../fieldset/__internal__/fieldset.context";
 
 export interface TextInputProps
   extends Omit<ValidationProps, "info">,
@@ -170,11 +171,21 @@ export const TextInput = React.forwardRef(
     });
     const hintId = useRef(guid());
     const inputHintId = inputHint ? hintId.current : undefined;
+    const inputPrefixId = prefix ? `${uniqueId}-prefix` : undefined;
+
+    const {
+      size: fieldsetSize,
+      hasError: fieldsetError,
+      required: fieldsetRequired,
+    } = useContext(FieldsetContext);
+    const actualSize = fieldsetSize || size;
+    const hasError = !!error || !!fieldsetError;
 
     const resolvedInputWidth =
       inputWidth ?? (labelInline ? INLINE_INPUT_WIDTH : INPUT_WIDTH);
     const ariaDescribedByString = [
       inputHintId,
+      inputPrefixId,
       ariaDescribedBy,
       ariaDescribedByProp,
     ]
@@ -187,7 +198,7 @@ export const TextInput = React.forwardRef(
     const validationMessage = hasValidationFailure && !disableErrorBorder && (
       <ValidationMessage
         id={validationId}
-        size={size}
+        size={actualSize}
         error={error}
         warning={warning}
       />
@@ -218,7 +229,7 @@ export const TextInput = React.forwardRef(
         data-element={dataElement}
         data-role={dataRole}
         data-component={dataComponent}
-        $size={size}
+        $size={actualSize}
         $labelInline={labelInline}
         $hasValidationFailure={labelInline && hasValidationFailure}
         {...filterStyledSystemMarginProps(rest)}
@@ -227,7 +238,7 @@ export const TextInput = React.forwardRef(
           <LabelWrapper
             data-role="label-wrapper"
             $labelInline={labelInline}
-            $size={size}
+            $size={actualSize}
             $labelWrapperWidth={
               labelInline && typeof resolvedInputWidth === "number"
                 ? 100 - resolvedInputWidth
@@ -237,14 +248,14 @@ export const TextInput = React.forwardRef(
             <Label
               id={labelId}
               htmlFor={uniqueId}
-              size={size}
+              size={actualSize}
               isRequired={required}
               {...stateProps}
             >
               {label}
             </Label>
             {inputHint && (
-              <HintText id={inputHintId} size={size} {...stateProps}>
+              <HintText id={inputHintId} size={actualSize} {...stateProps}>
                 {inputHint}
               </HintText>
             )}
@@ -252,7 +263,7 @@ export const TextInput = React.forwardRef(
         )}
         <InputWrapper
           data-role="input-wrapper"
-          $size={size}
+          $size={actualSize}
           $maxWidth={maxWidth}
           $inputWidth={inputWidth}
           onClick={() => {
@@ -264,29 +275,30 @@ export const TextInput = React.forwardRef(
               localRef.current?.focus();
             }
           }}
-          data-is-open={dataIsOpen}
         >
           {validationMessagePositionTop && validationMessage}
           <Input
             id={uniqueId}
             name={uniqueName}
-            aria-invalid={!!error}
+            aria-invalid={hasError}
             aria-describedby={ariaDescribedByString}
             aria-labelledby={ariaLabelledBy}
             value={value}
             placeholder={!(disabled || readOnly) ? placeholder : undefined}
-            required={required}
+            required={required || fieldsetRequired}
             ref={combinedRef}
             autoFocus={autoFocus}
             onMouseDown={handleMouseDown}
             onClick={handleClick}
             onChange={onChange}
-            error={!!error}
+            error={hasError}
             inputIcon={inputIcon}
-            size={size}
+            size={actualSize}
             prefix={prefix}
+            prefixId={inputPrefixId}
             leftChildren={leftChildren}
             type={type}
+            data-is-open={dataIsOpen}
             {...stateProps}
             {...filterOutStyledSystemSpacingProps(rest)}
           >