+++ src/components/textbox/__internal__/__next__/text-input.style.ts
@@ -55,6 +55,10 @@ const StyledTextInput = styled.div.attrs(applyBaseTheme)<StyledTextInputProps>`
     }
   }};
 
+  .fieldset-content & {
+    gap: var(--global-space-comp-xs);
+  }
+
   .time & {
     gap: unset;
   }
@@ -115,10 +119,6 @@ const InputWrapper = styled.div<InputWrapperProps>`
     ${$maxWidth && `max-width: ${$maxWidth};`}
     width: ${$inputWidth ? `${$inputWidth}%` : "100%"};
   `}
-
-  &[data-is-open="true"] {
-    z-index: var(--adaptiveSidebarModalBackdrop, 9999);
-  }
 `;
 
 export { StyledTextInput, LabelWrapper, InputWrapper };